tests: Dedicated Shader Debug Info file
